added smooth scroll
